+++ resources/lang/cs/translations.php
@@ -4,7 +4,7 @@ return [
     'breadcrumb' => 'Monitor úloh ve frontě',
     'title' => 'Úlohy ve frontě',
     'navigation_label' => 'Úlohy',
-    'navigation_group' => 'Systém',
+    'navigation_group' => 'Nastavení',
     'total_jobs' => 'Celkový počet zpracovaných úloh',
     'execution_time' => 'Celková doba provádění',
     'average_time' => 'Průměrná doba provádění',